Completed lab1
Task: Write a program for the Android platform that has an interface for entering and/or selection of data and displays the result of interaction with thisinterface in some text field of this interface. Provide the following: if not
all data is not entered or selected, and the user clicks the button to get the result,then display a pop-up window with a message to complete the input of all data.
The window must contain a collapsed list (phone types), a group of options (company), i.e. radio bars, and the OK button. Display information about the selection when clicking on the "OK" button in a text field.